fix #18 `codebase` was not always correctly initialized and used
