+++ src/main/java/de/theminefighter/stslauncher/JavaUtilities.java
@@ -1,18 +1,13 @@
 package de.theminefighter.stslauncher;
 
 import java.io.File;
-import java.net.MalformedURLException;
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.net.URL;
 import java.nio.file.Paths;
 
 /**
  * Some Java utility methods
  */
 public class JavaUtilities {
-	public static URI codebase;
-
 	/**
 	 * Determines the path of the java executable running
 	 *
@@ -43,14 +38,4 @@ public class JavaUtilities {
 		return !getStsLauncherLocation().isDirectory();
 	}
 
-	public static URL resolveHref(String href) {
-		try {
-			if (codebase == null) {
-				return new URL(href);
-			}
-			return codebase.resolve(href).toURL();
-		} catch (MalformedURLException e) {
-			throw new RuntimeException(e);
-		}
-	}
 }
