+++ src/main/java/javax/jnlp/impl/JWSContext.java
@@ -13,12 +13,15 @@ import javax.xml.parsers.ParserConfigurationException;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 
 public class JWSContext {
 	private static Element root;
 	private static Element information;
 	private static ResourceCache cache;
+	private static URI codebase;
 
 	private JWSContext() {}
 
@@ -45,6 +48,13 @@ public class JWSContext {
 			throw new RuntimeException("Error whilst determining jnlp context using jnlp file " + jnlp, e);
 		}
 		root = parse.getDocumentElement();
+		if (root.hasAttribute("codebase")) {
+			try {
+				codebase = new URI(root.getAttribute("codebase"));
+			} catch (URISyntaxException e) {
+				throw new RuntimeException("JNLP code base URI is invalid", e);
+			}
+		}
 		information = (Element) root.getElementsByTagName("information").item(0);
 	}
 
@@ -64,11 +74,11 @@ public class JWSContext {
 		NodeList iconTags = getInformation().getElementsByTagName("icon");
 		if (iconTags.getLength() == 0) return null;
 		String href = ((Element) iconTags.item(0)).getAttribute("href");
-		return cache.get(JavaUtilities.resolveHref(href), true);
+		return cache.get(resolveHref(href), true);
 	}
 
 	public static File getLocalJNLP() {
-		return cache.get(JavaUtilities.resolveHref(getRoot().getAttribute("href")), true);
+		return cache.get(resolveHref(getRoot().getAttribute("href")), true);
 
 	}
 
@@ -90,4 +100,14 @@ public class JWSContext {
 
 	}
 
+	public static URL resolveHref(String href) {
+		try {
+			if (codebase == null) {
+				return new URL(href);
+			}
+			return codebase.resolve(href).toURL();
+		} catch (MalformedURLException e) {
+			throw new RuntimeException(e);
+		}
+	}
 }
